update lab 7

+++ Lab7/src/academy/pocu/comp3500/lab7/Decryptor.java
@@ -37,7 +37,7 @@ public class Decryptor {
     private int getSumAscii(String str) {
         int asciiSum = 0;
         for (int i = 0; i < str.length(); i++) {
-            asciiSum += (int) str.charAt(i);
+            asciiSum += str.charAt(i);
         }
         return asciiSum;
     }
